+++ imm-20200930/alibabacloud_imm20200930/models/_dataset_config.py
@@ -9,12 +9,16 @@ class DatasetConfig(DaraModel):
     def __init__(
         self,
         insights: main_models.InsightsConfig = None,
+        smart_cluster: main_models.SmartClusterConfig = None,
     ):
         self.insights = insights
+        self.smart_cluster = smart_cluster
 
     def validate(self):
         if self.insights:
             self.insights.validate()
+        if self.smart_cluster:
+            self.smart_cluster.validate()
 
     def to_map(self):
         result = dict()
@@ -24,6 +28,9 @@ class DatasetConfig(DaraModel):
         if self.insights is not None:
             result['Insights'] = self.insights.to_map()
 
+        if self.smart_cluster is not None:
+            result['SmartCluster'] = self.smart_cluster.to_map()
+
         return result
 
     def from_map(self, m: dict = None):
@@ -31,6 +38,10 @@ class DatasetConfig(DaraModel):
         if m.get('Insights') is not None:
             temp_model = main_models.InsightsConfig()
             self.insights = temp_model.from_map(m.get('Insights'))
+
+        if m.get('SmartCluster') is not None:
+            temp_model = main_models.SmartClusterConfig()
+            self.smart_cluster = temp_model.from_map(m.get('SmartCluster'))
 
         return self